refactor!: remove opentelemetry
tl;dr it doesn't work well with blazor server

Since Blazor Server is primarily ran in the
browser, opentelemetry struggles to get useful
data to us. It really only shows that:

1. /_blazor is used a lot (duh)
2. connections are held open a long time (duh)
3. latency for page loading is 40,000+ ms (oof)

All of that paints a misleading picture of the
state of the application due to it's browser
centric workings. It works well in blazor wasm
but to get a good picture of blazor server it
needs something client side like hotjar.